refactor: move allDecksCounts to anki-common
Used by widgets and reminders - extraction needed for multi-module
support

Part of 20737
